extract diverging scales
